moving controller into game
