--- libs/controller/enums.ts
@@ -1,8 +1,0 @@
-const KEY_UP = 2048;
-const KEY_DOWN = 2049;
-const INTERNAL_KEY_UP = 2050;
-const INTERNAL_KEY_DOWN = 2051;
-const SYSTEM_KEY_UP = 2052;
-const SYSTEM_KEY_DOWN = 2053;
-const KEY_REPEAT = 2054;
-const SYSTEM_KEY_REPEAT = 2055;
