--- libs/controller/pxt.json
@@ -1,15 +1,0 @@
-{
-    "name": "controller",
-    "description": "Game controller",
-    "files": [
-        "enums.ts",
-        "controller.ts",
-        "multi.ts",
-        "targetoverrides.ts"
-    ],
-    "dependencies": {
-        "core": "file:../core",
-        "screen": "file:../screen"
-    },
-    "public": true
-}
