+++ libs/game/controller.ts
@@ -1,3 +1,340 @@
+const KEY_UP = 2048;
+const KEY_DOWN = 2049;
+const INTERNAL_KEY_UP = 2050;
+const INTERNAL_KEY_DOWN = 2051;
+const SYSTEM_KEY_UP = 2052;
+const SYSTEM_KEY_DOWN = 2053;
+const KEY_REPEAT = 2054;
+const SYSTEM_KEY_REPEAT = 2055;
+
+enum ControllerButtonEvent {
+    //% block="pressed"
+    Pressed = KEY_DOWN,
+    //% block="released"
+    Released = KEY_UP,
+    //% block="repeat"
+    Repeated = KEY_REPEAT
+}
+
+/**
+ * Access to game controls
+ */
+//% weight=98 color="#e15f41" icon="\uf11b"
+//% groups='["Single Player", "Multi Player"]'
+namespace controller {
+    let _userEventsEnabled = true;
+
+    //% fixedInstances
+    export class Button {
+        public id: number;
+        public repeatDelay: number;
+        public repeatInterval: number;
+        private _pressed: boolean;
+        private _pressedElasped: number;
+        private _repeatCount: number;
+        private _initEvents: () => void;
+
+        constructor(id: number, buttonId?: number, upid?: number, downid?: number) {
+            this.id = id;
+            this._pressed = false;
+            this.repeatDelay = 500;
+            this.repeatInterval = 30;
+            this._repeatCount = 0;
+            this._initEvents = () => {
+                control.internalOnEvent(INTERNAL_KEY_UP, this.id, () => {
+                    if (this._pressed) {
+                        this._pressed = false
+                        this.raiseButtonUp();
+                    }
+                }, 16)
+                control.internalOnEvent(INTERNAL_KEY_DOWN, this.id, () => {
+                    if (!this._pressed) {
+                        this._pressed = true;
+                        this._pressedElasped = 0;
+                        this._repeatCount = 0;
+                        this.raiseButtonDown();
+                    }
+                }, 16)
+                if (buttonId && upid && downid) {
+                    control.internalOnEvent(buttonId, upid, () => control.raiseEvent(INTERNAL_KEY_UP, this.id), 16)
+                    control.internalOnEvent(buttonId, downid, () => control.raiseEvent(INTERNAL_KEY_DOWN, this.id), 16)
+                }
+            }
+        }
+
+        initEvents() {
+            if (this._initEvents) {
+                this._initEvents();
+                this._initEvents = undefined;
+            }
+        }
+
+        private raiseButtonUp() {
+            if (_userEventsEnabled)
+                control.raiseEvent(KEY_UP, this.id)
+            else
+                control.raiseEvent(SYSTEM_KEY_UP, this.id);
+        }
+
+        private raiseButtonDown() {
+            if (_userEventsEnabled)
+                control.raiseEvent(KEY_DOWN, this.id)
+            else
+                control.raiseEvent(SYSTEM_KEY_DOWN, this.id)
+        }
+
+        private raiseButtonRepeat() {
+            if (_userEventsEnabled)
+                control.raiseEvent(KEY_REPEAT, this.id)
+            else
+                control.raiseEvent(SYSTEM_KEY_REPEAT, this.id)
+        }
+
+        /**
+         * Run some code when a button is pressed, released, or held
+         */
+        //% weight=99 blockGap=8 help=controller/button/on-event
+        //% blockId=keyonevent block="on %button **button** %event"
+        //% group="Single Player"
+        onEvent(event: ControllerButtonEvent, handler: () => void) {
+            this.initEvents();
+            control.onEvent(event, this.id, handler);
+        }
+
+        /**
+         * Pauses until a button is pressed or released
+         */
+        //% weight=98 blockGap=8 help=controller/button/pause-until
+        // blockId=keypauseuntil block="pause until %button **button** is %event"
+        //% group="Single Player"
+        pauseUntil(event: ControllerButtonEvent) {
+            this.initEvents();
+            control.waitForEvent(event, this.id)
+        }
+
+        /**
+         * Indicates if the button is currently pressed
+        */
+        //% weight=96 blockGap=8 help=controller/button/is-pressed
+        //% blockId=keyispressed block="is %button **button** pressed"
+        //% group="Single Player"
+        isPressed() {
+            return this._pressed;
+        }
+
+        setPressed(pressed: boolean) {
+            if (this._pressed != pressed) {
+                this._pressed = pressed;
+                if (this._pressed)
+                    this.raiseButtonUp();
+                else
+                    this.raiseButtonDown();
+            }
+        }
+
+        __update(dtms: number) {
+            if (!this._pressed) return;
+            this._pressedElasped += dtms;
+            // inital delay
+            if (this._pressedElasped < this.repeatDelay)
+                return;
+
+            // do we have enough time to repeat
+            const count = Math.floor((this._pressedElasped - this.repeatDelay) / this.repeatInterval);
+            if (count != this._repeatCount) {
+                this.raiseButtonRepeat();
+                this._repeatCount = count;
+            }
+        }
+    }
+
+    /**
+     * Pause the program until a button is pressed
+     */
+    //% weight=10
+    export function pauseUntilAnyButtonIsPressed() {
+        control.waitForEvent(KEY_DOWN, 0)
+    }
+
+    export function _setUserEventsEnabled(enabled: boolean) {
+        _userEventsEnabled = enabled;
+    }
+}
+
+enum ControllerButton {
+    //% block="left"
+    Left = 0,
+    //% block="up"
+    Up,
+    //% block="right"
+    Right,
+    //% block="down"
+    Down,
+    //% block="A"
+    A,
+    //% block="B"
+    B,
+    //% block="menu"
+    Menu
+}
+
+namespace controller {
+    let controllers: Controller[] = [];
+
+    //% fixedInstances
+    export class Controller {
+        private buttons: Button[];
+        // array of left,up,right,down,a,b,menu buttons
+        constructor(leftId: number) {
+            this.buttons = [];
+            for (let i = 0; i < 7; ++i)
+                this.buttons.push(new Button(leftId + i));
+
+            controllers.push(this);
+        }
+
+        /**
+         * Gets the left button
+         */
+        //%
+        get left() {
+            return this.button(ControllerButton.Left);
+        }
+
+        /**
+         * Gets the right button
+         */
+        //%
+        get right() {
+            return this.button(ControllerButton.Right);
+        }
+
+        /**
+         * Gets the right button
+         */
+        //%
+        get up() {
+            return this.button(ControllerButton.Up);
+        }
+
+        /**
+         * Gets the right button
+         */
+        //%
+        get down() {
+            return this.button(ControllerButton.Down);
+        }
+
+        /**
+         * Gets the right button
+         */
+        //%
+        get A() {
+            return this.button(ControllerButton.A);
+        }
+
+        /**
+         * Gets the right button
+         */
+        //%
+        get B() {
+            return this.button(ControllerButton.B);
+        }
+
+        /**
+         * Gets the right button
+         */
+        //%
+        get menu() {
+            return this.button(ControllerButton.Menu);
+        }
+
+        private button(button: ControllerButton): Button {
+            return this.buttons[button];
+        }
+
+        /**
+         * Run some code when a button is pressed, released, or held
+         */
+        //% weight=99 blockGap=8 help=controller/button/on-event
+        //% blockId=ctrlonevent block="on %controller %button **button** %event"
+        //% group="Multi Player"
+        onEvent(btn: ControllerButton, event: ControllerButtonEvent, handler: () => void) {
+            this.button(btn).onEvent(event, handler);
+        }
+
+        /**
+         * Indicates if the button is currently pressed
+        */
+        //% weight=96 blockGap=8 help=controller/button/is-pressed
+        //% blockId=ctrlispressed block="is %controller %button **button** pressed"
+        //% group="Multi Player"
+        isPressed(btn: ControllerButton): boolean {
+            return this.button(btn).isPressed();
+        }
+
+        /**
+         * Get the horizontal movement, given the step and state of buttons
+         * @param step the distance, eg: 100
+         */
+        //% weight=50 blockGap=8 help=controller/dx
+        //% blockId=ctrldx block="%controller dx (left-right buttons)||scaled by %step"
+        //% step.defl=100
+        //% group="Multi Player"
+        dx(step: number = 100) {
+            const ctx = control.eventContext();
+            if (!ctx) return 0;
+
+            if (this.left.isPressed()) {
+                if (this.right.isPressed()) return 0
+                else return -step * ctx.deltaTime;
+            }
+            else if (this.right.isPressed()) return step * ctx.deltaTime
+            else return 0
+        }
+
+        /**
+         * Get the vertical movement, given the step and state of buttons
+         * @param step the distance, eg: 100
+         */
+        //% weight=49 help=keys/dy
+        //% blockId=ctrldy block="%controller dy (up-down buttons)||scaled by %step"
+        //% step.defl=100
+        //% group="Multi Player"
+        dy(step: number = 100) {
+            const ctx = control.eventContext();
+            if (!ctx) return 0;
+
+            if (this.up.isPressed()) {
+                if (this.down.isPressed()) return 0
+                else return -step * ctx.deltaTime;
+            }
+            else if (this.down.isPressed()) return step * ctx.deltaTime
+            else return 0
+        }
+
+        __update(dt: number) {
+            const dtms = (dt * 1000) | 0
+            this.buttons.forEach(btn => btn.__update(dtms));
+        }    
+    }
+
+    /**
+     * Called by the game engine to update and/or raise events
+     */
+    export function __update(dt: number) {
+        const dtms = (dt * 1000) | 0
+        controllers.forEach(ctrl => ctrl.__update(dtms));
+    }
+
+    export function serialize(offset: number): Buffer {
+        const buf = control.createBuffer(offset + 1);
+        return buf;
+    }
+
+    export function onEvent()
+}
+
 namespace controller {
     interface ControlledSprite {
         s: Sprite;
