+++ libs/game/docs/reference/controller/move-sprite.md
